requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
SQLAlchemy==2.0.30
python-dotenv==0.21.0